refactor(dsh-provider-usage): 阶段四——apply/接口面收口（#670）
Summary of changes
- 新增 src/apply/interface.ts 目录门面（最小面具名导出：ReportConfigService 类型）
- apply/apply.ts 与 apply/index.ts 内指向 domain1/domain2/shared 的直引全部
  收口到各层 interface.ts（含 domain2 全部 7 层：aggregate/collect/common/
  execute/schedule/routes 与既有 domain1 各层/shared）
- domain2/routes/reports.ts 对 apply/report-config-service.ts 的直引改走
  apply/interface.ts（apply 目录现在有唯一对外门面）
- apply/index.ts 作为 lib 入口保持既有导出面不变（213 键守恒，仅改 import 来源）

Design decisions and rationale
- interface.ts 是目录门面，服务目录外源码引用；lib 入口仍是 apply/index.ts
  （bundle-host 锚点）。apply()/inject/ROUTES 属 lib 契约面，当前无目录外
  源码引用，不入 interface.ts（避免无消费者的镜像导出）
- apply/interface.ts 最小面仅导出 ReportConfigService：它是 apply/ 目录目前
  被目录外（domain2/routes/reports.ts）消费的唯一符号
- B 梯队确认各层 interface.ts 已覆盖 apply 所需全部符号，本提交零新增导出，
  未触碰 src/domain1/**、src/domain2/**（reports.ts 仅改 1 行 import）、
  src/shared/** 内容

Related issue/ticket references
- wingsky-1/dsh-plugin-hub#670（阶段四 C1）
- #664 D10 目录门面决策
