+++ packages/dsh-provider-usage/src/apply/apply.ts
@@ -37,22 +37,22 @@ import { HotReloadableAdapter } from "../domain1/registry/interface.ts";
 import { resolvePath } from "../domain1/registry/interface.ts";
 import { Config, normalizeConfig, type NormalizedConfig } from "../shared/interface.ts";
 import { readUiConfig } from "../shared/interface.ts";
-import { makeLayerErrorSurface } from "../domain2/common/errsurf.ts";
+import { makeLayerErrorSurface } from "../domain2/common/interface.ts";
 import { readAdapterStateResult, readUserAdapters } from "../domain1/registry/interface.ts";
 import { loadUserAdapterChecked } from "../domain1/registry/interface.ts";
 import { StatsServiceCtor as StatsService } from "../domain1/pipeline/interface.ts";
-import { TrendTracker } from "../domain2/aggregate/index.ts";
-import { readReportConfig, type ReportConfig } from "../domain2/schedule/config.ts";
-import { ReportScheduler } from "../domain2/schedule/scheduler.ts";
-import { optionalNotifier } from "../domain2/execute/runner.ts";
+import { TrendTracker } from "../domain2/aggregate/interface.ts";
+import { readReportConfig, type ReportConfig } from "../domain2/schedule/interface.ts";
+import { ReportScheduler } from "../domain2/schedule/interface.ts";
+import { optionalNotifier } from "../domain2/execute/interface.ts";
 import { ReportConfigService } from "./report-config-service.ts";
-import { makeDueReportExecutor } from "../domain2/execute/executor.ts";
-import { makeListDirs } from "../domain2/execute/list-dirs.ts";
-import { ReportTaskQueue } from "../domain2/schedule/tasks.ts";
+import { makeDueReportExecutor } from "../domain2/execute/interface.ts";
+import { makeListDirs } from "../domain2/execute/interface.ts";
+import { ReportTaskQueue } from "../domain2/schedule/interface.ts";
 import { createStatsRoutes } from "../domain1/routes/interface.ts";
 import { createAdapterRoutes } from "../domain1/routes/interface.ts";
-import { createUiRoutes } from "../domain2/routes/ui.ts";
-import { createReportRoutes } from "../domain2/routes/reports.ts";
+import { createUiRoutes } from "../domain2/routes/interface.ts";
+import { createReportRoutes } from "../domain2/routes/interface.ts";
 import type {} from "@deepseek-ai/dsh-session";
 
 export const ROUTES: Record<string, string> = {
